pyproject.toml
[project]
name = "cdx-trade"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alpaca-py>=0.43.2",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "python-dotenv>=1.0.1",
]
